pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Jarvis-like voice assistant with local vLLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "vllm>=0.11.2",
    "openai-whisper>=20231117",
    "piper-tts>=1.2.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
